refactor: decompose App.tsx into AppRouter component, useScrollFade hook, normalizeProjects utility, and move View type to types.ts
