+++ landing-page/src/hooks/useScrollFade.ts
@@ -1,0 +1,42 @@
+import { useEffect } from "react";
+import type { View } from "../types";
+
+export function useScrollFade(view: View) {
+  useEffect(() => {
+    if (view !== "landing" && view !== "about") {
+      document
+        .querySelectorAll(".fade-target")
+        .forEach((el) => el.classList.add("visible"));
+      return;
+    }
+
+    let timer: ReturnType<typeof setTimeout> | undefined;
+    if (view === "landing") {
+      const heroTargets = document.querySelectorAll(".hero-fade");
+      timer = setTimeout(() => {
+        heroTargets.forEach((el) => el.classList.add("visible"));
+      }, 100);
+    }
+
+    const observer = new IntersectionObserver(
+      (entries) => {
+        entries.forEach((entry) => {
+          if (entry.isIntersecting) {
+            entry.target.classList.add("visible");
+            observer.unobserve(entry.target);
+          }
+        });
+      },
+      { threshold: 0.1 },
+    );
+
+    document
+      .querySelectorAll(".scroll-fade")
+      .forEach((el) => observer.observe(el));
+
+    return () => {
+      if (timer) clearTimeout(timer);
+      observer.disconnect();
+    };
+  }, [view]);
+}
